Refactor: 부스 상품 이미지 서비스 분리
